perf(codec)!: read a whole JSON number as a Long
Every number in a parsed document used to be a BigDecimal. That is
exact, which was the point — the document model before it parsed
numbers as Double and silently rounded anything above 2^53 — but it
is also two objects and about sixty bytes, spent on a value that is
nearly always a row id, a count, or a timestamp offset. A Forgejo
response is mostly numbers, so the whole decode path paid it.

WHAT CHANGES

JsonValue has two number cases now instead of one:

  JsonValue.Int64(value: Long)          a number written without a
                                        fractional part or an
                                        exponent whose digits fit in
                                        64 bits
  JsonValue.Decimal(value: BigDecimal)  everything else: fractional,
                                        exponent form, or a whole
                                        number past Long.MaxValue

Which case a parsed number lands in follows the text that was on the
wire and nothing else, so Json.render still writes back exactly what
it read — 102 stays 102 and 102.0 stays 102.0. Forgejo's integer
fields reject 102.0, so that was never negotiable; the existing
round-trip tests pinning it are untouched and green.

The parse calls jsoniter's readNumber rather than readBigDecimal.
readNumber returns a java.lang.Long for a whole number that fits, a
java.math.BigInteger for one that does not, and a
java.math.BigDecimal for everything else — checked against the 2.39.1
jar rather than recalled — which is exactly the distinction the two
cases want to draw.

PRECISION IS UNCHANGED

Nothing here goes near a Double. A Long holds every integer up to
2^63 exactly and anything larger stays a BigDecimal, so
9007199254740993 (2^53 + 1, the first integer a Double cannot
represent) and 9223372036854775808 (one past Long.MaxValue) both
round-trip byte for byte. There is a test for each.

MEASURED, NOT ASSUMED

scripts/alloc-bench.sh, full table, before and after, on OpenJDK
64-Bit Server VM 25.0.4+7-LTS. B/op is heap bytes allocated per
operation.

  operation              before       after    change
  ------------------------------------------------------
  parse.page-50         904,128     889,328     -1.6%
  decode.page-50      1,159,616   1,134,416     -2.2%
  decode.repo-wide       23,200      22,680     -2.2%
  assemble.repo-wide      5,096       4,888     -4.1%
  decode.org-narrow       3,160       3,096     -2.0%
  assemble.org-narrow       608         592     -2.6%
  parse.ints-1000        88,528      56,952    -35.7%
  parse.bools-1000       26,576      26,560      noise
  baseline.noop             0.0         0.0         —

parse.bools-1000 is the control: this change cannot touch an array of
booleans, and it did not move beyond the 16 bytes the harness header
calls noise.

parse.ints-1000 parses [0,1,...,999], whose numbers are far smaller
than a real row id, so a row-id-shaped array was added to the harness
temporarily, measured on both sides of the change with the same
filter, and then removed — a thousand nine-digit integers, the shape
Codeberg's identifiers actually have:

  parse.ids-1000         95,080      63,080    -33.7%

That is 32.0 bytes saved per number. It is also why the page rows
move so little: repo-single.json holds 13 numbers against 44 strings
and 35 booleans, so numbers were never most of a repository's weight.
They are most of a page of identifiers or a quota response.

WHAT IS LEFT ON THE TABLE

63.1 bytes an element is still not the 26.6 the same-length array of
booleans costs. About 24 of the difference is a java.lang.Long that
jsoniter boxes in order to return it from readNumber, and that dies
one line later. Removing it needs a look-ahead the reader interface
does not offer; the only route to it is to try readLong behind a mark
and catch the failure, which hands a remote party a document that
throws an exception once per fractional number. Twenty-four bytes is
the cheaper of the two, so the box stays, documented where it
happens.

SOURCE COMPATIBILITY

JsonValue.Num survives as an object with the same four constructors
and a new unapply, so JsonValue.Num(7) still builds a number and
`case JsonValue.Num(value)` still binds a BigDecimal. What it can no
longer be is a type.

The readers on the hot path were moved off it rather than left on it,
because the extractor has to build a BigDecimal for the Int64 case
and that is the cost this commit exists to remove. JsonDecoder.long
and JsonFields.number name the two cases instead, through the new
JsonValue.longOpt; RepositoryContentDto asked numOpt.isDefined and
now asks the new JsonValue.isNum, which allocates nothing; and
EditorConfigDto's renderer matches the cases directly.

TWO CORRECTNESS FIXES ALONG THE WAY

BranchProtectionOptionDto and MigrateRepoOptionsDto built their
numbers as JsonValue.Num(x.toDouble), which rounds above 2^53 and so
could have corrupted a team id — an int64 on the wire. They pass the
value itself now. EditorConfigDto rendered a whole number through
BigDecimal.toLong, which wraps rather than fails for a value too
large for a Long, and uses toBigInt now.

VERIFICATION

./verify.sh --with-slow passes. Duplication is unchanged at 373
groups, so the recorded baseline is left where it is.

BREAKING CHANGE: JsonValue.Num is no longer a case class. A number in
a parsed document is now JsonValue.Int64(Long) or
JsonValue.Decimal(BigDecimal), and JsonValue.Num is an object holding
the constructors and an extractor for both.

Constructing is unchanged: JsonValue.Num(7), Num(7L), Num(7.0) and
Num(BigDecimal(7)) all still compile. They return JsonValue rather
than JsonValue.Num, which no longer names a type, and Num is the
constructor to keep using — it is what decides which of the two cases
a value belongs in, and building a JsonValue.Decimal directly from a
whole number that would fit a Long produces a document that does not
equal the one parsing its own rendering gives back.

Matching with `case JsonValue.Num(value)` is unchanged and still
binds a BigDecimal, but it now builds one for a whole number. Prefer
`case JsonValue.Int64(value)` where a Long was what was wanted, or
the new JsonValue.longOpt.

A match that enumerates every JsonValue case must replace its
`case JsonValue.Num(_)` with the two new cases. The compiler says so:
Num is no longer one of the cases, so such a match stops being
exhaustive.

+++ modules/codec/src/com/worxbend/codeberg4s/repositories/admin/wire/MigrateRepoOptionsDto.scala
@@ -95,7 +95,7 @@ private[codeberg4s] object TransferRepoOptionDto:
     List(
       Some(NewOwnerKey -> JsonValue.Str(command.newOwner.value)),
       Option.when(command.teamIds.nonEmpty)(
-        TeamIdsKey     -> JsonValue.Arr.from(command.teamIds.map(team => JsonValue.Num(team.value.toDouble)))
+        TeamIdsKey     -> JsonValue.Arr.from(command.teamIds.map(team => JsonValue.Num(team.value)))
       ),
     ).flatten
